move event-db types and quries to the legacy mod
