+++ catalyst-gateway/bin/src/event_db/legacy/mod.rs
@@ -1,0 +1,4 @@
+//! Legacy event db queries and types
+
+pub(crate) mod queries;
+pub(crate) mod types;
