+++ catalyst-gateway/bin/src/event_db/legacy/queries/mod.rs
