+++ catalyst-gateway/bin/src/event_db/legacy/queries/vit_ss/mod.rs
