+++ catalyst-gateway/bin/src/event_db/legacy/types/ballot.rs
@@ -1,6 +1,6 @@
 //! Ballot types
 use super::{objective::ObjectiveId, proposal::ProposalId};
-use crate::event_db::types::registration::VoterGroupId;
+use crate::event_db::legacy::types::registration::VoterGroupId;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 /// Objective Choices
