+++ catalyst-gateway/bin/src/event_db/legacy/types/objective.rs
@@ -1,7 +1,7 @@
 //! Objective Types
 use serde_json::Value;
 
-use crate::event_db::types::registration::VoterGroupId;
+use crate::event_db::legacy::types::registration::VoterGroupId;
 
 #[allow(clippy::module_name_repetitions)]
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
