+++ catalyst-gateway/bin/src/event_db/legacy/types/search.rs
