+++ catalyst-gateway/bin/src/event_db/legacy/types/vit_ss/vote_plan.rs
