+++ catalyst-gateway/bin/src/event_db/legacy/types/voting_status.rs
