+++ catalyst-gateway/bin/src/service/api/legacy/v1/fragments_post.rs
@@ -4,8 +4,8 @@ use poem_extensions::{response, UniResponse::T200};
 use poem_openapi::payload::Json;
 
 use crate::service::common::{
-    objects::{
-        legacy::fragments_batch::FragmentsBatch, legacy::fragments_processing_summary::FragmentsProcessingSummary,
+    objects::legacy::{
+        fragments_batch::FragmentsBatch, fragments_processing_summary::FragmentsProcessingSummary,
     },
     responses::{
         resp_2xx::OK,
