+++ catalyst-gateway/bin/src/service/common/objects/legacy/event_id.rs
@@ -13,8 +13,8 @@ impl Example for EventId {
     }
 }
 
-impl From<EventId> for crate::event_db::types::event::EventId {
+impl From<EventId> for crate::event_db::legacy::types::event::EventId {
     fn from(event_id: EventId) -> Self {
-        crate::event_db::types::event::EventId(event_id.0)
+        crate::event_db::legacy::types::event::EventId(event_id.0)
     }
 }
