+++ catalyst-gateway/bin/src/service/common/objects/legacy/voter_info.rs
@@ -54,11 +54,11 @@ impl Example for VoterInfo {
     }
 }
 
-impl TryFrom<crate::event_db::types::registration::VoterInfo> for VoterInfo {
+impl TryFrom<crate::event_db::legacy::types::registration::VoterInfo> for VoterInfo {
     type Error = String;
 
     fn try_from(
-        value: crate::event_db::types::registration::VoterInfo,
+        value: crate::event_db::legacy::types::registration::VoterInfo,
     ) -> Result<Self, Self::Error> {
         Ok(Self {
             voting_power: value.voting_power,
