+++ catalyst-gateway/bin/src/service/common/objects/legacy/voter_registration.rs
@@ -34,10 +34,12 @@ impl Example for VoterRegistration {
     }
 }
 
-impl TryFrom<crate::event_db::types::registration::Voter> for VoterRegistration {
+impl TryFrom<crate::event_db::legacy::types::registration::Voter> for VoterRegistration {
     type Error = String;
 
-    fn try_from(value: crate::event_db::types::registration::Voter) -> Result<Self, Self::Error> {
+    fn try_from(
+        value: crate::event_db::legacy::types::registration::Voter,
+    ) -> Result<Self, Self::Error> {
         Ok(Self {
             voter_info: value.info.try_into()?,
             as_at: value.as_at,
